[AIMode Attribution] Enable Explore page to inject a highlight chart and support multiple places. (#5114)
This PR uses the URL Hash Params to add the ability to pass in multiple
places, and pass in one statistical variable.
We inject the chart for the one stat var, and issue a second call to
fetch the Topic of said SV, then issue another fulfill call to with the
associated topic.

Adds support for multiple StatVars, and multiple Topics on the fulfill
call.
Does a very basic merging of the responses.

In this state, here's what the page looks like, while passing in:
* Rhode Island and Arizona
* Median_Income_Household, Area_Farm, Count_Person
* TIMELINE_WITH_HIGHLIGHT for chartType


http://localhost:8080/explore/#sv=Median_Income_Household___Area_Farm___Count_Person&p=geoId%2F48___geoId%2F06&client=ui_related_topic&chartType=TIMELINE_WITH_HIGHLIGHT


https://screencast.googleplex.com/cast/NDg2NDE3MjA1ODczODY4OHwyODllMGExYS0wMQ

Note that the first 3 charts are for each of the StatVars passed in. The
remaining come from the 3 topics identified for the statVars
dc/topic/Income, dc/topic/Agriculture, dc/topic/Demographics

Next Up: Allow injecting provenance, unit and timerange. more Code
Cleanup. Testing.