Metadata Feature Flag (#5157)
## Description

This PR creates a new feature flag to mediate access to the new metadata
modal.

The feature flag, `metadata_modal` is set to false in all environments
except for `local` and `autopush`.

If the feature flag is set to false, the modal will display the older
metadata content (which contains links to stat vars).

### Additional Update

In testing this, I noticed that now that we've moved the older
stat-var-based metadata modal to the new standard modal, the flicker on
load was a little more noticeable (the modal wasn't handling the loading
state explicitly when it fetched the information it needed).

To improve this, I've made a minor update to the default (stat var)
modal, `tile_metadata_modal_simple`, to wire its loading state into the
dialog so that it displays a proper loading state.

## Testing

To test this, first keep the flags as they are, and go to a page such as
this:


http://localhost:8080/tools/visualization#visType%3Dtimeline%26place%3Dcountry%2FSWE%26placeType%3DAdministrativeArea1%26sv%3D%7B%22dcid%22%3A%22LifeExpectancy_Person%22%7D

The metadata modal should display the new, updated metadata.

Then, flip the `local` version of the `metadata_modal` flag to false.
You may need to then restart the server to have it catch the updated
flag, but once it does, open the modal, and you will see the older modal
that only shows stat vars.