[Chart Injection] Make the highlight Response only show the highlighted chart + Remove the matching chart from the non-highlight response. (#5165)
This PR does two main things for improving the chart injection work:

For single chart injection, it would show the proper chart + a chart for
the contained in places. This PR removes that part of the code. Whenever
we have that chartType URL parameter, we filter out any additional
contained_in logic for chart generation.

For all charts, since we duplicated the call to fulfill while
restricting to a single stat var, that chart will appear in the
highlight chart + the regular response. so I implemented the logic to
remove any duplicate blocks from the non-highlight block.

Examples:
Before: https://screenshot.googleplex.com/BtvrNX3hHEGaBjv
After: https://screenshot.googleplex.com/9WF3EDSQLVtEZn9
Note the contained in block is gone, and the duplicate first chart is
gone.