[Production Excellence] Filter out deprecated flags from test (#5184)
This test checks that we never have a flag that is enabled in production
but missing in the other files.

In this PR I just filter out the features that are marked as deprecated
such that we can delete it without modifying the production.json file at
the same time.