Highlight Metadata Loading (#5188)
## Description

This PR adds the following functionality to the highlight result
metadata loading:

* The component now keeps a loading state, and feeds that loading state
through to the MetadataSummary component (moved out to its own component
now that we've extended this part of the functionality.
* The loading state shows a spinner and loading message (using the
shared loader) and then once loaded, displays the metadata. By having
the loading, we also prevent a page bump to the components below once
the metadata summary has loaded.
* We were previously using the `block.description` to pass the metadata
summary down through to the block. I was worried that overloading the
description might interfere with actual block descriptions, so I created
an extension to the block that adds `metadataSummary`. I didn't add it
directly into the interface in `subject_page_proto_types.ts` because
this interface represents the shape of what the API returns (and there
is a plan to have that autogenerated from the API).
* We now send the metadataSummary through using this attribute. This
way, we also can clearly separate metadata from description for the
purposes of the loading state.
* The fetch now no longer occurs when there is no highlightFacet (this
was triggering a runtime error error - although one that did not affect
the display of the page - when there was no highlight facet).

We will not display an error directly to the user if the metadata fetch
fails (but there is a TODO for reporting this error through Google
Analytics).

## Testing

The following link will demonstrate the loading state (and then display
the metadata as before):


http://localhost:8080/explore#sv=LifeExpectancy_Person&p=country%2FITA___country/FRA&client=ui_related_topic&chartType=TIMELINE_WITH_HIGHLIGHT&unit=Euro&imp=WorldDevelopmentIndicators

This link shows a highlight result that does not have a highlight facet.
This will display as before (without a loading state or any metadata and
the runtime error now gone).


http://localhost:8080/explore#sv=Count_Person&p=country%2FFRA___country%2FITA&client=ui_related_topic&chartType=TIMELINE_WITH_HIGHLIGHT