Add altSentences metadata generation for Stat Vars using Gemini API (#5191)
Adds a function in the python script `add_metadata.py` to call the
Gemini API using existing stat var metadata to generate alternative
sentences. Tested by running locally and verifying the change.

Adding this function makes the script very slow - It took approximately
80 minutes to process every single stat var. In a follow up PR I will
try to make the Gemini calls to different batches in parallel, which
will hopefully improve performance. Also still need to modify the export
so that it will be compatible with exporting to cloud storage once this
is run as a cloud function.