Facet Selector Adjustments (#5214)
## Description

This PR makes the following adjustments to the layout and copy of the
facet selector:

- "Close" has become "Cancel" in the facet selector modal.
- The "action" button in the modals now appear on the right. In other
words, "Update" appears to the right of "Cancel". This will be the
standard going forward.
- The facet selector is now rounded and has a smaller font to better
match its surroundings in the chart.

## Notes

The font size here is a special case (to match the design of the
existing charts) but is otherwise not part of the current design
standard. Therefore we applied it specifically to these components only
rather than adding it to the theme.

## Screenshots

### Modal buttons

<img width="639" alt="Screenshot 2025-06-16 at 1 36 37 PM"
src="https://github.com/user-attachments/assets/6b7337a7-2d31-4dde-928b-e4db74b8e2b9"
/>

### Facet selector update

<img width="845" alt="Screenshot 2025-06-16 at 2 13 39 PM"
src="https://github.com/user-attachments/assets/7bbc9698-797a-4df0-84a2-8b197d91b23c"
/>

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>