Dataset Selector Modal V2 (#5221)
## Description

This PR implements the new facet selector dialog with an enriched
metadata. This corresponds to V2 of the dialog from the slide deck
(located on [page
16](https://docs.google.com/presentation/d/1ztyZwu8b3YzU1MMAFPI3iC-Hm00CE63rPb7eP6vGZlw/edit?pli=1&slide=id.g347cc273f35_0_120#slide=id.g347cc273f35_0_120)
of the slide-deck).

Because each chart in each vis tool collects and prepares the facets to
be sent into the facet selector independently (and in different ways
across the different charts and vis tools), this PR touches all the
different charts across the two versions of the vis tools. However, we
made efforts to isolate the changes as much as possible to the aspects
of each chart that supplies the facet selector (in other words, this PR
should not affect data sent into the charts themselves).

## Notes

### Architectural of Facet Selector Promises

Before this PR, the facet selector component took in a promise, rather
than a list of facets, and then resolved the promise itself. The old vis
tools and download tool passed the facet list in as an already resolved
promise (with an empty facet list before the fetch had occurred). The
new vis tools handled in a promise that was not yet resolved.

These variations complicate handling of loading states. A note that this
was always an issue; however it wasn’t noticeable until we introduced
the number of facet counts into the button, i.e., `Select a dataset
[3]`. The old facet selector just displayed `Edit Sources` even as the
facets were still loading, effectively relying on the loading to have
finished by the time the user clicked the button.

The new facet selector needs to know when it is loading so that we only
display the button along with the facet count once loading is complete.
This is more robust in general.

To do this we wanted to standardize how the information was transmitted
and convert it to a more common React pattern. The facet selector now
take a list and loading state rather than a promise.

We introduced a custom hook to resolve those promises in the new vis
tools. The older tools were structured in such a way that this was not
necessary.

### Measuring methods

As discussed, the measuring methods display without a label (i.e., no
"Measuring method: " if we have a full description of the measuring
method). If we

## Screenshots

### Before

<img width="623" alt="before"
src="https://github.com/user-attachments/assets/06c52886-31f0-499e-91b8-1cb89c95cc9f"
/>

### After

<img width="615" alt="after 3"
src="https://github.com/user-attachments/assets/82b0a88f-7702-422a-a3de-4dedea4f480c"
/>

## Feature Flag

The functionality that displays the V2 (enriched-metadata) facet
selector dialog is hidden behind the same feature flag that mediates the
display of the new rich metadata modal.

Once that flag is removed and V2 because live, we will remove the old
facet selector entirely (this marked as a TODO). In the meantime, the
new facet selector is called FacetSelectorRich, the old one
FacetSelectorSimple, and FacetSelector uses the feature flag to
determine which to display.

## Tests

Because the feature flag causes the facet selector modal to render
differently in cdc versus standard charts during testing, the
`select_source` functionality in the shared test suite is updated to
take multiple possible options to search for, allowing us to provide
search options for both the old and the new versions of the selector at
the same time in the test.

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>