[Productionization Excellence] Make NL Integration Test goldens automatic from Cloud Build (#5298)
This creates a Cloud Build Action which uses our DataCommons Robot
Author in order to automatically update goldens and send us a PR to
review with all the diffs.

It takes around 9 minutes to run for now. 
The idea is that this will be part of the automated flow after releasing
to staging.

See sample PR (note I was running it on a different branch to show some
diffs!) - https://github.com/datacommonsorg/website/pull/5299

Also removes the attempt to do this via cloud build.