Add "filled" icons to generation script (#5327)
Adds the ability to create a "filled" icon using the icon generation
script.

Adds the flags `--include_filled` and `--filled_only` to generate an
icon with the [fill
axis](https://developers.google.com/fonts/docs/material_symbols#fill_axis)
set to 1.

Changes added as followup to PR #5294 

---
Testing strategy:

Run the script with
```
./run.sh change_circle --include_filled
./run.sh favorite --filled_only
./run.sh bolt --include_filled --flask
./run.sh star --filled_only --react
```
and see that the proper files are generated