Add source filtering for Vertex AI (#5373)
Website-side follow up from
https://github.com/datacommonsorg/mixer/pull/1585.

Modifies the Vertex AI pathway in `search_statvar` to make a call to
`Mixer.FilterStatVarsByEntity` to implement source filtering for stat
var searches done through Vertex AI. As part of this change, I modified
the initial limit for searches from 100 to 300 if filtering by sources
to act as a buffer for the filter.

Tested locally to make sure that stat var explorer, map, and timeline
all still work. Also modified the Python tests - I need to add more
tests specifically testing that the source selection is working, but
when fixing the existing tests that also verified that
`dc.filter_statvars` is being called when it should be.

---------

Co-authored-by: Gabriel Mechali <gabriel.mechali@gmail.com>