Small fixes to release process that broke (#5405)
I updated the cache clearing to be a cloud build action. The wget
command had failed on it. This PR will fix it by making it use curl.

On the CustomDC script update, this seemed to be needed as the script
kept failing. Either the github CLI version on my own repo is higher, or
they deprecated the old format. With this change, I was able to create
the PR to update CDC.