Facet Selector - Scroll to selected option (#5456)
## Description

With this PR, when a user opens the facet selector after already having
selected an option in that same facet selector previously, the facet
selector will automatically appear scrolled with that item centered in
view.

This is row number 5. in the
[Bugbash](https://docs.google.com/spreadsheets/d/1VtmR1H3VOBbEL_D0OxP9z0ZSFmD3H6k_ZldraM11CjA/edit?gid=690212228#gid=690212228),
"Would be nice that when you open the data selector dialog it goes to
the place where the data is selected, vs just the top".

## Notes

I tried a version first where it opened scrolled to the top, and have it
automatically scrolled down smoothly to the selected item, but this
turned out to be a bit "dizzying". Now, we rather use a useLayoutEffect
to actually scroll to that item before the user sees anything, and the
effect is nicer.

Another note is that the diff is fairly large but a lot of that is the
result of increased indentation because of the forwardRef.

## Testing

A convenient page test this is:

http://localhost:8080/explore#q=Population+of+the+world

The "Population in the World" chart has a large number of options,
allowing the scrolling to be tested easily.